Move _decode_private_key_content function to SnowflakeBaseProfileMapping
Move function to SnowflakeBaseProfileMapping to avoid duplicating code on
SnowflakePrivateKeyPemProfileMapping and SnowflakeEncryptedPrivateKeyPemProfileMapping
