+++ cosmos/profiles/snowflake/user_privatekey.py
@@ -2,8 +2,6 @@
 
 from __future__ import annotations
 
-import base64
-import binascii
 import json
 from typing import TYPE_CHECKING, Any
 
@@ -44,28 +42,6 @@ class SnowflakePrivateKeyPemProfileMapping(SnowflakeBaseProfileMapping):
         "role": "extra.role",
         "private_key": "extra.private_key_content",
     }
-
-    def _decode_private_key_content(self, private_key_content: str) -> str:
-        """
-        Decodes the private key content from either base64-encoded or plain-text PEM format.
-
-        Starting from `apache-airflow-providers-snowflake` version 6.3.0, the provider expects the
-        `private_key_content` to be base64-encoded rather than raw PEM text. This method ensures
-        compatibility by attempting to decode the content from base64 first. If decoding fails,
-        the original content is assumed to be plain-text PEM (as used in older versions).
-
-        This allows backward compatibility while supporting the new expected format.
-
-        Args:
-            private_key_content: The private key content, either base64 encoded or plain-text PEM
-
-        Returns:
-            The decoded private key in plain-text PEM format
-        """
-        try:
-            return base64.b64decode(private_key_content).decode("utf-8")
-        except (UnicodeDecodeError, binascii.Error):
-            return private_key_content
 
     @property
     def conn(self) -> Connection:
